requirements.txt
fastapi[all]==0.111.0
apscheduler==3.10.4
sqlalchemy==2.0.35
psycopg==3.2.1
psycopg-binary==3.2.1
requests==2.32.3
python-dateutil==2.9.0
